Revert Jackson upgrade and manually resolve it in SBT, and clean up Hive hack
