+++ dev/deps/spark-deps-hadoop-3.1
@@ -88,17 +88,15 @@ htrace-core4-4.1.0-incubating.jar
 httpclient-4.5.4.jar
 httpcore-4.4.8.jar
 ivy-2.4.0.jar
-jackson-annotations-2.8.11.jar
-jackson-core-2.8.11.jar
+jackson-annotations-2.6.7.jar
+jackson-core-2.6.7.jar
 jackson-core-asl-1.9.13.jar
-jackson-databind-2.8.11.jar
-jackson-dataformat-yaml-2.8.11.jar
-jackson-jaxrs-base-2.7.8.jar
-jackson-jaxrs-json-provider-2.7.8.jar
+jackson-databind-2.6.7.1.jar
+jackson-dataformat-yaml-2.6.7.jar
 jackson-mapper-asl-1.9.13.jar
-jackson-module-jaxb-annotations-2.8.11.jar
-jackson-module-paranamer-2.8.11.jar
-jackson-module-scala_2.11-2.8.11.jar
+jackson-module-jaxb-annotations-2.6.7.jar
+jackson-module-paranamer-2.7.9.jar
+jackson-module-scala_2.11-2.6.7.1.jar
 janino-3.0.8.jar
 java-xmlbuilder-1.1.jar
 javassist-3.18.1-GA.jar
@@ -200,7 +198,7 @@ scala-xml_2.11-1.0.5.jar
 shapeless_2.11-2.3.2.jar
 slf4j-api-1.7.16.jar
 slf4j-log4j12-1.7.16.jar
-snakeyaml-1.17.jar
+snakeyaml-1.15.jar
 snappy-0.2.jar
 snappy-java-1.1.7.1.jar
 spire-macros_2.11-0.13.0.jar
