+++ sql/hive-thriftserver/src/test/scala/org/apache/spark/sql/hive/thriftserver/HiveThriftServer2Suites.scala
@@ -808,6 +808,8 @@ abstract class HiveThriftJdbcTest extends HiveThriftServer2Test {
 }
 
 abstract class HiveThriftServer2Test extends SparkFunSuite with BeforeAndAfterAll with Logging {
+  Utils.hiveShimsHack()
+
   def mode: ServerMode.Value
 
   private val CLASS_NAME = HiveThriftServer2.getClass.getCanonicalName.stripSuffix("$")
